+++ src/Execution/DataLoader/PaginatedRelationLoader.php
@@ -70,7 +70,7 @@ class PaginatedRelationLoader implements RelationLoader
 
                 $relation->initRelation([$model], $relationName);
 
-                /** @var \Illuminate\Database\Eloquent\Relations\Relation&\Illuminate\Database\Eloquent\Builder $relation */
+                // @phpstan-ignore-next-line Builder mixin is not understood
                 return $relation->forPage($this->paginationArgs->page, $this->paginationArgs->first);
             });