+++ src/Schema/Directives/ComplexityDirective.php
@@ -3,7 +3,6 @@
 namespace Nuwave\Lighthouse\Schema\Directives;
 
 use Closure;
-use Illuminate\Support\Arr;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Utils;
@@ -27,28 +26,28 @@ directive @complexity(
 GRAPHQL;
     }
 
-    public function handleField(FieldValue $fieldFieldValue, Closure $next): FieldValue
+    public function handleField(FieldValue $fieldValue, Closure $next): FieldValue
     {
         if ($this->directiveHasArgument('resolver')) {
             [$className, $methodName] = $this->getMethodArgumentParts('resolver');
 
             $namespacedClassName = $this->namespaceClassName(
                 $className,
-                $fieldFieldValue->defaultNamespacesForParent()
+                $fieldValue->defaultNamespacesForParent()
             );
 
             $resolver = Utils::constructResolver($namespacedClassName, $methodName);
         } else {
-            $resolver = function (int $childrenComplexity, array $args): int {
+            $resolver = static function (int $childrenComplexity, array $args): int {
                 /** @var int $complexity */
-                $complexity = Arr::get($args, 'first', 1);
+                $complexity = $args['first'] ?? 1;
 
                 return $childrenComplexity * $complexity;
             };
         }
 
-        return $next(
-            $fieldFieldValue->setComplexity($resolver)
-        );
+        $fieldValue->setComplexity($resolver);
+
+        return $next($fieldValue);
     }
 }