+++ src/SoftDeletes/TrashedDirective.php
@@ -4,10 +4,8 @@ namespace Nuwave\Lighthouse\SoftDeletes;
 
 use Exception;
 use GraphQL\Exception\InvalidArgument;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 use Laravel\Scout\Builder as ScoutBuilder;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Scout\ScoutBuilderDirective;
@@ -41,13 +39,16 @@ GRAPHQL;
             return $builder;
         }
 
-        /** @var Builder&SoftDeletes $builder */
+        /** @see \Illuminate\Database\Eloquent\SoftDeletes */
         switch ($value) {
             case 'with':
+                // @phpstan-ignore-next-line because it involves mixins
                 return $builder->withTrashed();
             case 'only':
+                // @phpstan-ignore-next-line because it involves mixins
                 return $builder->onlyTrashed();
             case 'without':
+                // @phpstan-ignore-next-line because it involves mixins
                 return $builder->withoutTrashed();
             default:
                 throw new InvalidArgument('Unexpected value for Trashed filter: '.$value);