+++ tests/database/migrations/2018_02_28_000004_create_testbench_posts_table.php
@@ -6,9 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateTestbenchPostsTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('posts', function (Blueprint $table): void {
@@ -31,9 +28,6 @@ class CreateTestbenchPostsTable extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::drop('posts');