+++ tests/database/migrations/2018_12_08_000001_create_testbench_categories_table.php
@@ -6,22 +6,20 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateTestbenchCategoriesTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('categories', function (Blueprint $table): void {
             $table->increments('category_id');
             $table->string('name');
 
+            $table->unsignedInteger('parent_id')->nullable();
+
             $table->timestamps();
+
+            $table->foreign('parent_id')->references('category_id')->on('categories');
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::drop('categories');